- Changed the internal directory structure to load classes properly.
